chore: rename 'uiSchema' in PluginConfig to 'formSchema' (#2634)
* PluginConfig.uiSchema ->formSchema

* fix alphabetical issue in test
