+++ Composer/packages/extensions/adaptive-form/src/PluginContext.ts
@@ -5,7 +5,7 @@ import React from 'react';
 import { PluginConfig } from '@bfc/extension';
 
 const PluginContext = React.createContext<Required<PluginConfig>>({
-  uiSchema: {},
+  formSchema: {},
   visualSchema: {},
   roleSchema: {},
   // kindSchema: {},
