+++ Composer/packages/extensions/extension/src/types/extension.ts
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { UISchema, RoleSchema, RecognizerSchema } from './uiSchema';
+import { UISchema, RoleSchema, RecognizerSchema } from './formSchema';
 import { FlowEditorConfig } from './flowSchema';
 
 export interface PluginConfig {
-  uiSchema?: UISchema;
+  formSchema?: UISchema;
   roleSchema?: RoleSchema;
   recognizers?: RecognizerSchema[];
   visualSchema?: FlowEditorConfig;
